Move and rename UnreadWidgetConfigurationActivity
