+++ k9mail/src/main/AndroidManifest.xml
@@ -285,7 +285,7 @@
             android:name=".activity.EmailAddressList"
             android:configChanges="locale"/>
 
-        <activity android:name=".activity.UnreadWidgetConfiguration">
+        <activity android:name=".widget.unread.UnreadWidgetConfigurationActivity">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_CONFIGURE"/>
             </intent-filter>
