+++ k9mail/src/main/res/xml/unread_widget_info.xml
@@ -4,6 +4,6 @@
     android:previewImage="@drawable/preview_unread_widget"
     android:minHeight="40dp"
     android:minWidth="40dp"
-    android:configure="com.fsck.k9.activity.UnreadWidgetConfiguration"
+    android:configure="com.fsck.k9.widget.unread.UnreadWidgetConfigurationActivity"
     android:updatePeriodMillis="0">
 </appwidget-provider>
